Rewrite _write() for block writes and incomplete checks
usbd_ep_write_packet() returns the number of bytes successfully written.
When the USB peripheral is busy writing out a packet, it will return 0,
as no bytes can be written.

As a result, the previous version of _write() would usually fail to
write all but the first character, due to too-fast calls to
usbd_ep_write_packet().

A quick workaround is simply to make a single call to in _write(), for
the entire string.  However, quick calls to _write() could still cause
the same issue.

Ideally, the return value from usbd_ep_write_packet() could simply be
returned directly from _write(), as callers must be able to handle
incomplete writes.  Unfortunately, while Newlib can handle incomplete
writes, it can only do so when at least one byte is written.  It's
fflush() (which is eventually called from puts()) returns in error if
_write() returns 0 [1].

Thus, the complete retry loop is placed within _write(), which does not
return until the entire string is written.  Note that there is no
timeout mechanism in place currently.

This also drops the functionality of replacing \n with \r on write, but
it is simple enough to manually add \r to strings, if desired.

[1] https://github.com/aeste/newlib/blob/fd857b01b85d3ce3ec9763b71acaf753d030e596/newlib/libc/stdio/fflush.c#L214-215